Revert "Changes and some errors resolved"
