+++ app/src/main/AndroidManifest.xml
@@ -6,7 +6,6 @@
     <uses-permission android:name="android.permission.INTERNET" />
     <uses-permission android:name="android.permission.READ_EXTERNAL_STORAGE" />
     <uses-permission android:name="android.permission.WRITE_EXTERNAL_STORAGE" />
-    <uses-permission android:name="android.permission.MEDIA_CONTENT_CONTROL" />
 
     <application
         android:name=".MyApplication"
@@ -236,14 +235,6 @@
         <activity
             android:name=".SelectSubject"
             android:label="@string/title_activity_select_subject" >
-        </activity>
-        <activity
-            android:name=".UploadImageMultiple"
-            android:label="@string/title_activity_upload_image_multiple" >
-        </activity>
-        <activity
-            android:name=".CustomPhotoGalleryActivity"
-            android:label="@string/title_activity_custom_photo_gallery" >
         </activity>
     </application>
