+++ app/src/main/res/layout/activity_login.xml
@@ -83,8 +83,6 @@
         android:layout_centerVertical="true"
         android:layout_alignParentEnd="true"
         android:layout_alignParentRight="true"
-
-
         android:background="#20BF9F"
         android:textColor="#ffffff" />
