+++ app/src/main/res/layout/activity_new_class.xml
@@ -1,5 +1,9 @@
 <?xml version="1.0" encoding="utf-8"?>
-
+<ScrollView xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="fill_parent"
+    android:layout_height="fill_parent"
+    android:fillViewport="true">
 <RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
     android:orientation="vertical" android:layout_width="match_parent"
@@ -32,20 +36,19 @@
         android:text="New Grade:"
         android:id="@+id/textView36"
         android:layout_marginTop="33dp"
-        android:textSize="20sp"
+        android:textSize="25sp"
         android:textColor="#ffff"
         android:layout_below="@+id/container_toolbar"
         android:layout_alignParentStart="true" />
 
     <EditText
-        android:layout_width="300sp"
-        android:layout_height="wrap_content"
+        android:layout_width="300dp"
+        android:layout_height="30dp"
         android:id="@+id/newClassGrade"
         android:layout_below="@+id/textView36"
         android:layout_alignParentLeft="true"
         android:layout_alignParentStart="true"
-        android:backgroundTint="#FFFFFF"
-        android:textColor="#ffffff"/>
+        android:backgroundTint="#FFFFFF" />
 
     <TextView
         android:layout_width="wrap_content"
@@ -53,20 +56,19 @@
         android:textAppearance="?android:attr/textAppearanceMedium"
         android:text="Section Name:"
         android:id="@+id/rollno"
-        android:textSize="20sp"
+        android:textSize="25sp"
         android:textColor="#ffff"
         android:layout_below="@+id/newClassGrade"
         android:layout_alignParentStart="true"
         android:layout_marginTop="38dp" />
 
     <EditText
-        android:layout_width="300sp"
-        android:layout_height="wrap_content"
+        android:layout_width="300dp"
+        android:layout_height="30dp"
         android:id="@+id/newClassSection"
         android:layout_below="@+id/rollno"
         android:layout_alignParentStart="true"
-        android:backgroundTint="#FFFFFF"
-        android:textColor="#ffffff"/>
+        android:backgroundTint="#FFFFFF" />
 
     <TextView
         android:layout_width="wrap_content"
@@ -74,7 +76,7 @@
         android:textAppearance="?android:attr/textAppearanceMedium"
         android:text="Class Teacher:"
         android:id="@+id/textview"
-        android:textSize="20sp"
+        android:textSize="25sp"
         android:textColor="#ffff"
         android:layout_below="@+id/newClassSection"
         android:layout_alignParentStart="true"
@@ -107,20 +109,20 @@
         android:textAppearance="?android:attr/textAppearanceMedium"
         android:text="Class Teacher Subject:"
         android:id="@+id/subjecttxt"
-        android:textSize="20sp"
+        android:textSize="25sp"
         android:textColor="#ffff"
         android:layout_marginTop="32dp"
         android:layout_below="@+id/classteacherselection"
         android:layout_alignParentStart="true" />
 
     <EditText
-        android:layout_width="300sp"
-        android:layout_height="wrap_content"
+        android:layout_width="300dp"
+        android:layout_height="30dp"
         android:id="@+id/classteachersubject"
         android:layout_below="@+id/subjecttxt"
         android:layout_alignParentStart="true"
-        android:backgroundTint="#FFFFFF"
-        android:textColor="#ffffff"/>
+        android:backgroundTint="#FFFFFF" />
 
 
 </RelativeLayout>
+    </ScrollView>
