+++ app/src/main/res/layout/activity_new_teacher.xml
@@ -31,7 +31,7 @@
         android:text="Name:"
         android:id="@+id/textView36"
 
-        android:textSize="20sp"
+        android:textSize="25sp"
         android:textColor="#ffff"
         android:layout_marginTop="70dp"
         android:layout_below="@+id/container_toolbar"
@@ -64,7 +64,7 @@
         android:layout_height="wrap_content"
         android:text="Age:"
         android:id="@+id/textView8"
-        android:textSize="20sp"
+        android:textSize="25sp"
         android:textColor="#ffffff"
         android:layout_centerVertical="true"
         android:layout_alignParentStart="true" />
@@ -86,7 +86,7 @@
         android:id="@+id/textView9"
 
         android:textColor="#ffffff"
-        android:textSize="20sp"
+        android:textSize="25sp"
         android:layout_marginTop="48dp"
         android:layout_below="@+id/teacherAge"
         android:layout_alignParentStart="true" />
