+++ app/src/main/res/layout/activity_sign_up.xml
@@ -42,8 +42,6 @@
         android:id="@+id/nameText"
         android:layout_marginTop="33dp"
         android:layout_alignParentTop="true"
-        android:layout_marginStart="15dp"
-        android:layout_alignParentStart="true"
         android:layout_toStartOf="@+id/passwordSignup" />
 
     <TextView
@@ -58,8 +56,6 @@
         android:id="@+id/emailText"
         android:layout_marginTop="31dp"
         android:layout_below="@+id/nameText"
-        android:layout_marginStart="15dp"
-        android:layout_alignParentStart="true"
         android:layout_alignEnd="@+id/nameText" />
 
     <TextView
@@ -73,9 +69,7 @@
         android:scrollHorizontally="true"
         android:id="@+id/passwordText"
         android:layout_marginTop="31dp"
-        android:layout_marginStart="15dp"
         android:layout_below="@+id/emailText"
-        android:layout_alignParentStart="true"
         android:layout_alignEnd="@+id/emailText" />
 
     <TextView
@@ -89,9 +83,7 @@
         android:scrollHorizontally="true"
         android:id="@+id/confirmPasswordText"
         android:layout_marginTop="31dp"
-        android:layout_marginStart="15dp"
         android:layout_below="@+id/passwordText"
-        android:layout_alignParentStart="true"
         android:layout_alignEnd="@+id/passwordText"
          />
 
@@ -101,7 +93,6 @@
         android:inputType="textPersonName"
         android:ems="10"
         android:id="@+id/userNameSignup"
-        android:layout_toRightOf="@+id/nameText"
         android:layout_alignBottom="@+id/nameText"
         android:layout_toEndOf="@+id/nameText" />
 
@@ -111,7 +102,6 @@
         android:inputType="textEmailAddress"
         android:ems="10"
         android:id="@+id/emailSignup"
-        android:layout_toRightOf="@+id/emailText"
         android:layout_alignBottom="@+id/emailText"
         android:layout_toEndOf="@+id/emailText" />
 
@@ -122,7 +112,6 @@
         android:ems="10"
         android:id="@+id/passwordSignup"
         android:layout_alignBottom="@+id/passwordText"
-
         android:layout_alignEnd="@+id/already" />
 
     <EditText
@@ -131,9 +120,9 @@
         android:inputType="textPassword"
         android:ems="10"
         android:id="@+id/confirmPasswordSignup"
-
-        android:layout_alignBottom="@+id/confirmPasswordText"
-        android:layout_toEndOf="@+id/confirmPasswordText" />
+        android:layout_above="@+id/signUpButton"
+        android:layout_alignParentRight="true"
+        android:layout_alignParentEnd="true" />
 
     <TextView
         android:layout_width="wrap_content"
@@ -141,8 +130,6 @@
         android:textAppearance="?android:attr/textAppearanceSmall"
         android:text="Already a SmartEdu User?"
         android:id="@+id/already"
-
-
         android:layout_alignParentBottom="true"
         android:layout_alignParentEnd="true" />
 
@@ -152,13 +139,12 @@
         android:text="Sign Up"
         android:id="@+id/signUpButton"
 
-        android:layout_marginEnd="15dp"
+        android:layout_marginTop="78dp"
+        android:layout_below="@+id/confirmPasswordText"
+        android:layout_alignRight="@+id/passwordSignup"
+        android:layout_alignEnd="@+id/passwordSignup"
         android:background="#20BF9F"
-        android:textColor="#ffffff"
-        android:layout_below="@+id/confirmPasswordSignup"
-        android:layout_alignParentEnd="true"
-        android:layout_alignParentRight="true"
-        android:layout_alignEnd="@+id/passwordSignup" />
+        android:textColor="#ffffff" />
 
     </RelativeLayout>
