+++ app/src/main/res/layout/show_attendance.xml
@@ -14,7 +14,7 @@
         android:layout_height="wrap_content"
         android:text="Number of Absents"
         android:id="@+id/absentDays_head"
-        android:textSize="20sp"
+        android:textSize="25sp"
         android:layout_marginTop="57dp"
         android:layout_below="@+id/date"
         android:layout_alignParentLeft="true"
@@ -25,7 +25,7 @@
         android:layout_height="wrap_content"
         android:text="Total Days"
         android:id="@+id/totalDays_head"
-        android:textSize="20sp"
+        android:textSize="25sp"
         android:layout_alignParentLeft="true"
         android:layout_alignParentStart="true"
         android:layout_marginTop="25dp"
@@ -97,7 +97,7 @@
         android:layout_marginTop="25dp"
         android:layout_below="@+id/totalMarks"
         android:layout_alignParentLeft="true"
-        android:textSize="20sp" />
+        android:textSize="25sp" />
     <TextView
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
