+++ app/src/main/res/layout/show_upload_details.xml
@@ -1,6 +1,12 @@
+<ScrollView xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="fill_parent"
+    android:layout_height="fill_parent"
+    android:fillViewport="true">
+
 <RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools" android:layout_width="match_parent"
-    android:layout_height="match_parent" android:paddingLeft="@dimen/activity_horizontal_margin"
+    android:layout_height="wrap_content" android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
     android:paddingBottom="@dimen/activity_vertical_margin"
@@ -44,7 +50,7 @@
         android:backgroundTint="#fff"
         android:nestedScrollingEnabled="false"
         android:layout_alignBottom="@+id/date"
-        android:layout_alignParentLeft="true"
+
         android:layout_alignParentStart="true" />
 
 
@@ -132,40 +138,6 @@
         android:layout_below="@+id/subject"
         android:layout_centerHorizontal="true" />
 
-
-    <TextView
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="Topic/Details"
-        android:id="@+id/textView20"
-        android:textSize="18sp"
-        android:textColor="#ffffff"
-        android:layout_alignTop="@+id/typeHead"
-        android:layout_alignParentEnd="true" />
-
-    <TextView
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:id="@+id/topic"
-        android:text="topic/details"
-        android:textSize="15sp"
-        android:layout_below="@+id/textView20"
-        android:layout_alignRight="@+id/textView20"
-        android:layout_alignEnd="@+id/textView20"
-        android:textColor="#ffffff" />
-
-    <Button
-        style="?android:attr/buttonStyleSmall"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="View All"
-        android:id="@+id/viewAll"
-        android:textSize="11sp"
-        android:textColor="#ffff"
-        android:background="#040404"
-        android:layout_alignParentBottom="true"
-        android:layout_centerHorizontal="true" />
-
     <Button
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
@@ -187,8 +159,45 @@
         android:layout_alignParentBottom="true"
         android:layout_alignParentStart="true" />
 
+    <TextView
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="Topic/Details"
+        android:singleLine="true"
+        android:ellipsize="marquee"
+        android:lines="1"
+        android:scrollHorizontally="true"
+        android:id="@+id/textView20"
+        android:layout_alignTop="@+id/typeHead"
+        android:layout_alignRight="@+id/doneButton"
+        android:layout_alignEnd="@+id/doneButton"
+        android:textSize="18sp"
+        android:textColor="#ffffff" />
 
+    <TextView
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:id="@+id/topic"
+        android:text="topic/details"
+        android:textSize="15sp"
+        android:layout_below="@+id/textView20"
+        android:layout_alignLeft="@+id/doneButton"
+        android:layout_alignStart="@+id/doneButton"
+        android:textColor="#ffffff" />
+
+    <Button
+        style="?android:attr/buttonStyleSmall"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="View All"
+        android:id="@+id/viewAll"
+        android:textSize="11sp"
+        android:textColor="#ffff"
+        android:background="#040404"
+        android:layout_alignParentBottom="true"
+        android:layout_centerHorizontal="true" />
 
 
 </RelativeLayout>
 
+</ScrollView>
