+++ app/src/main/res/layout/show_upload_details_students.xml
@@ -139,9 +139,10 @@
         android:text="Topic/Details"
         android:id="@+id/textView20"
         android:textSize="18sp"
-        android:textColor="#ffffff"
         android:layout_alignTop="@+id/typeHead"
-        android:layout_alignParentEnd="true" />
+        android:layout_toRightOf="@+id/viewAll"
+        android:layout_toEndOf="@+id/viewAll"
+        android:textColor="#ffffff" />
 
     <TextView
         android:layout_width="wrap_content"
